backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy>=3.1.0
flask-jwt-extended>=4.6.0
bcrypt>=4.0.0
python-dotenv==1.1.0
admet-ai
openai>=1.0.0
python-docx>=1.0.0
gunicorn>=21.0.0
pytest>=7.0.0
